:recycle: model 삭제
