--- src/main/java/com/potato/ecommerce/domain/member/repository/MemberRepository.java
@@ -1,9 +1,0 @@
-package com.potato.ecommerce.domain.member.repository;
-
-import com.potato.ecommerce.domain.member.model.Member;
-
-public interface MemberRepository {
-    Member findMemberBy(String email);
-    void save(Member member);
-    void update(Member member);
-}
